refactor(folder-structure): organize shared components into categorized Shared folders and update imports
